pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "capi-etl"
version = "0.1.0"
description = "ETL pipeline: Jira + GitHub → PostgreSQL (BI)"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.32",
    "pandas>=2.2",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
]

[project.scripts]
capi-etl = "capi_etl.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/capi_etl"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.14",
]
